+++ cordapp/src/main/kotlin/net/corda/plugins/CordappPlugin.kt
@@ -5,6 +5,7 @@ import org.gradle.api.GradleException
 import org.gradle.api.InvalidUserDataException
 import org.gradle.api.Plugin
 import org.gradle.api.Project
+import org.gradle.api.*
 import org.gradle.api.artifacts.Configuration
 import org.gradle.api.artifacts.Dependency
 import org.gradle.jvm.tasks.Jar
@@ -76,16 +77,40 @@ class CordappPlugin : Plugin<Project> {
         }
         jarTask.dependsOn(task)
 
-        val signTask = project.task("signCordappJar")
+        val optionalSignTask = createSigningTask(project, "singCordapp", false)
+        jarTask.finalizedBy(optionalSignTask)
+
+        val signTask = project.tasks.findByName("sourceJar") //task in corda-publish-utils plugin
+        if (signTask != null ) {
+            val singTaskForPublishingTask = createSigningTask(project, "singCordappForPublishing", true)
+            val jarPublishingTask = signTask as Jar
+            jarPublishingTask.dependsOn(singTaskForPublishingTask)
+        }
+    }
+
+    private fun createSigningTask(project: Project, taskName: String, isEnableByDefault: Boolean) : Task {
+        val signTask = project.task(taskName)
         signTask.doLast {
-            if (cordapp.signing.enabled) {
+            var enabled = !isEnableByDefault && cordapp.signing.enabled == true  // Case when signing is optional by default
+            var enabledByDefault = isEnableByDefault && cordapp.signing.enabled != false  // Case when signing is required by default
+            if (enabled || enabledByDefault)  {
+                val path = project.tasks.getByName("jar").outputs.files.singleFile.toPath()
+                project.logger.debug( "Signing JAR " + path.fileName)
                 val options = cordapp.signing.options.toSignJarOptionsMap()
-                options["jar"] = project.tasks.getByName("jar").outputs.files.singleFile.toPath().toString()
-                project.ant.invokeMethod("signjar", options)
+                options["jar"] = path.toString()
+                try {
+                    // Any subsequent sign using the same key will be ignored
+                    project.ant.invokeMethod("signjar", options)
+                } catch (e: Exception) {
+                    throw InvalidUserDataException("Exception while signing ${path.fileName}," +
+                            " ensure the 'cordapp.signing.options' entry contains correct keyStore configuration," +
+                            " or disable signing by 'cordapp.signing.enabled=false'.", e)
+                }
             }
         }
-        jarTask.finalizedBy(signTask)
+        return signTask
     }
+
 
     private fun getDirectNonCordaDependencies(project: Project): Set<File> {
         project.logger.info("Finding direct non-corda dependencies for inclusion in CorDapp JAR")