+++ cordapp/src/main/kotlin/net/corda/plugins/Signing.kt
@@ -8,7 +8,8 @@ import org.gradle.util.ConfigureUtil
 open class Signing {
 
     @get:Input
-    var enabled: Boolean = false
+    // Depending on invoking task null value indicates the flag not being set and allows to treat it either as true or false by default.
+    var enabled: Boolean? = null
         private set
     fun enabled(value: Boolean) { enabled = value }